+++ inst/include/Rcpp/sugar/functional/functional.h
@@ -1,0 +1,6 @@
+#ifndef Rcpp__sugar__functional__functional_h
+#define Rcpp__sugar__functional__functional_h
+
+#include <Rcpp/sugar/functional/negate.h>
+
+#endif