+++ inst/include/Rcpp/sugar/sugar.h
@@ -4,6 +4,7 @@
 #include <Rcpp/sugar/tools/iterator.h>
 #include <Rcpp/sugar/block/block.h>
 
+#include <Rcpp/sugar/functional/functional.h>
 #include <Rcpp/sugar/operators/operators.h>
 #include <Rcpp/sugar/functions/functions.h>
 #include <Rcpp/sugar/lazy/lazy.h>